web_backEnd/src/ShopWeb.Tests/Service/AddressServiceTests.cs
using System;
using System.Collections.Generic;
using System.Threading.Tasks;
using AutoMapper;
using Moq;
using ShopWeb.Core.src.Entity;
using ShopWeb.Core.src.RepoAbstract;
using ShopWeb.Service.src.DTOs;
using ShopWeb.Service.src.Services;
using Xunit;

namespace ShopWeb.Service.Tests
{
    public class AddressServiceTests
    {
        private readonly AddressService _addressService;
        private readonly Mock<IAddressRepository> _addressRepositoryMock;
        private readonly Mock<IMapper> _mapperMock;

        public AddressServiceTests()
        {
            _addressRepositoryMock = new Mock<IAddressRepository>();
            _mapperMock = new Mock<IMapper>();

            _addressService = new AddressService(_addressRepositoryMock.Object, _mapperMock.Object);
        }

        [Fact]
        public async Task GetAllAddressesByUserIdAsync_ValidCustomerId_ReturnsAddressList()
        {
            // Arrange
            var customerId = Guid.NewGuid();
            var addresses = new List<Address>
            {
                // Set up list of addresses
            };
            _addressRepositoryMock.Setup(x => x.GetAllAddressesByUserIdAsync(customerId)).ReturnsAsync(addresses);
            _mapperMock.Setup(x => x.Map<IEnumerable<AddressReadDto>>(addresses)).Returns(new List<AddressReadDto>());

            // Act
            var result = await _addressService.GetAllAddressesByUserIdAsync(customerId);

            // Assert
            Assert.NotNull(result);
            // Add additional assertions as needed
        }

        [Fact]
        public async Task GetAddressByIdAsync_ValidAddressId_ReturnsAddress()
        {
            // Arrange
            var addressId = Guid.NewGuid();
            var address = new Address
            {
                // Set up address
            };
            _addressRepositoryMock.Setup(x => x.GetAddressByIdAsync(addressId)).ReturnsAsync(address);
            _mapperMock.Setup(x => x.Map<AddressReadDto>(address)).Returns(new AddressReadDto());

            // Act
            var result = await _addressService.GetAddressByIdAsync(addressId);

            // Assert
            Assert.NotNull(result);
            // Add additional assertions as needed
        }

        [Fact]
        public async Task CreateAddressAsync_ValidAddressDtoAndUserId_ReturnsCreatedAddress()
        {
            // Arrange
            var addressCreateDto = new AddressCreateDto
            {
                // Set up address create DTO
            };
            var userId = Guid.NewGuid();
            var address = new Address
            {
                // Set up address
            };

            _mapperMock.Setup(x => x.Map<Address>(addressCreateDto)).Returns(address);
            _addressRepositoryMock.Setup(x => x.CreateAddressAsync(address)).ReturnsAsync(address);
            _mapperMock.Setup(x => x.Map<AddressReadDto>(address)).Returns(new AddressReadDto());

            // Act
            var result = await _addressService.CreateAddressAsync(addressCreateDto, userId);

            // Assert
            Assert.NotNull(result);
            // Add additional assertions as needed
        }

        [Fact]
        public async Task UpdateAddressByIdAsync_ValidAddressDtoAndAddressId_ReturnsTrue()
        {
            // Arrange
            var addressId = Guid.NewGuid();
            var addressUpdateDto = new AddressUpdateDto
            {
                // Set up address update DTO
            };

            var updatedAddress = new Address
            {
                // Set up updated address
            };

            _mapperMock.Setup(x => x.Map<Address>(addressUpdateDto)).Returns(updatedAddress);
            _addressRepositoryMock.Setup(x => x.UpdateAddressByIdAsync(addressId, updatedAddress)).ReturnsAsync(true);

            // Act
            var result = await _addressService.UpdateAddressByIdAsync(addressUpdateDto, addressId);

            // Assert
            Assert.True(result);
            // Add additional assertions as needed
        }

        [Fact]
        public async Task DeleteAddressByIdAsync_ValidAddressId_ReturnsTrue()
        {
            // Arrange
            var addressId = Guid.NewGuid();

            _addressRepositoryMock.Setup(x => x.DeleteAddressByIdAsync(addressId)).ReturnsAsync(true);

            // Act
            var result = await _addressService.DeleteAddressByIdAsync(addressId);

            // Assert
            Assert.True(result);
        }
    }
}
